Change a sensors light level to int32
While the API documentation states the value is an int16 its value can be greater than the max int16.

+++ sensors/sensors.go
@@ -48,7 +48,7 @@ type State struct {
 	ButtonEvent int16  `json:"buttonevent,omitempty"`
 	Status      int16  `json:"status,omitempty"`
 	Temperature int16  `json:"temperature,omitempty"`
-	LightLevel  int16  `json:"lightlevel,omitempty"`
+	LightLevel  int32  `json:"lightlevel,omitempty"`
 	Dark        bool   `json:"dark,omitempty"`
 	Daylight    bool   `json:"daylight,omitempty"`
 }
